[simplify] refactor: extract shared git helpers, remove duplication
- Extract runGit/gitStdout/GitResult to src/utils/git.ts (was duplicated
  verbatim in trust-ledger.ts and reconciliation.ts)
- Replace private gitObjectExists in trust-ledger.ts with import from
  execution-state.ts where it already exists
- Tighten TrustRecord.status from bare string to ValidatorStatus | 'skipped'
